Unify storyboard-related swift files for iOS, tvOS and OSX.
The targets are switched by #if os().
