--- Swinject/iOS-tvOS/Container+Typealias.swift
@@ -1,14 +1,0 @@
-//
-//  Container+Typealias.swift
-//  Swinject
-//
-//  Created by Yoichi Tagaya on 8/1/15.
-//  Copyright © 2015 Swinject Contributors. All rights reserved.
-//
-
-import UIKit
-
-extension Container {
-    /// The typealias to UIViewController.
-    public typealias Controller = UIViewController
-}
