+++ Swinject/iOS-tvOS/UIStoryboard+Swizzling.swift
@@ -1,18 +1,29 @@
 //
-//  UIStoryboard+Swizzling.swift
+//  Storyboard+Swizzling.swift
 //  Swinject
 //
 //  Created by Yoichi Tagaya on 10/10/15.
 //  Copyright © 2015 Swinject Contributors. All rights reserved.
 //
 
-import UIKit
+#if os(iOS) || os(tvOS)
 
-extension UIStoryboard {
+private typealias Storyboard = UIStoryboard
+
+#elseif os(OSX)
+
+private typealias Storyboard = NSStoryboard
+
+#endif
+
+
+#if os(iOS) || os(tvOS) || os(OSX)
+
+extension Storyboard {
     // Class method `load` is not available in Swift. Instead `initialize` is used.
     // http://nshipster.com/swift-objc-runtime/
     public override class func initialize() {
-        if self !== UIStoryboard.self {
+        if self !== Storyboard.self {
             return
         }
         
@@ -20,16 +31,16 @@ extension UIStoryboard {
             static var token: dispatch_once_t = 0
         }
         dispatch_once(&Static.token) {
-            let original = class_getClassMethod(UIStoryboard.self, "storyboardWithName:bundle:")
-            let swizzled = class_getClassMethod(UIStoryboard.self, "swinject_storyboardWithName:bundle:")
+            let original = class_getClassMethod(Storyboard.self, "storyboardWithName:bundle:")
+            let swizzled = class_getClassMethod(Storyboard.self, "swinject_storyboardWithName:bundle:")
             method_exchangeImplementations(original, swizzled)
         }
     }
     
     @objc
-    private class func swinject_storyboardWithName(name: String, bundle storyboardBundleOrNil: NSBundle) -> UIStoryboard {
-        if self === UIStoryboard.self {
-            // Instantiate SwinjectStoryboard if UIStoryboard is tried to be instantiated.
+    private class func swinject_storyboardWithName(name: String, bundle storyboardBundleOrNil: NSBundle) -> Storyboard {
+        if self === Storyboard.self {
+            // Instantiate SwinjectStoryboard if UI/NSStoryboard is tried to be instantiated.
             return SwinjectStoryboard.create(name: name, bundle: storyboardBundleOrNil)
         } else {
             // Call original `storyboardWithName:bundle:` method swizzled with `swinject_storyboardWithName:bundle:`
@@ -38,3 +49,5 @@ extension UIStoryboard {
         }
     }
 }
+
+#endif
